chore(build) allow building as dynamic module
Adjust the config script so that `liblmdb.a` is linked first with
module.o, instead of nginx object. When building this module as dynamic
module, this fix ensures the statically linked library stays with the
module.so.

Also adding `-fPIC` to lmdb CFLAGS so when it statically linked, the
host (dynamic) library can be dynamically loaded.